==> task1_depression_symptoms/05_results_formatting.ipynb <==
import os
import pandas as pd

# Directorio donde están tus archivos csv
directorio = 'all-MiniLM-L12-v2'

# Obtener una lista de todos los archivos csv en el directorio
archivos = [os.path.join(directorio, f) for f in os.listdir(directorio)]

# Leer y concatenar todos los archivos csv
df = pd.concat((pd.read_csv(f) for f in archivos))

# Guardar el resultado en un nuevo archivo csv
df.to_csv('all-MiniLM-L12-v2/result.csv', index=False)

# Directorio donde están tus archivos csv
directorio = 'classifier 1'

# Obtener una lista de todos los archivos csv en el directorio
archivos = [os.path.join(directorio, f) for f in os.listdir(directorio)]

# Leer y concatenar todos los archivos csv
df = pd.concat((pd.read_csv(f) for f in archivos))

# Ordenar por 'label' y 'relevance' de manera descendente
df.sort_values(['label', 'relevance'], ascending=[True, False], inplace=True)

# Seleccionar los primeros 1000 registros de cada 'label'
df = df.groupby('label').head(1000)

# Guardar el nuevo dataframe en un csv
df.to_csv('classifier 1/dataset_final.csv', index=False)

df = pd.read_csv('submission/all-MiniLM-L6-v2.csv')
# Supongamos que 'df' es tu dataframe final
df = df.groupby('label').apply(lambda x: x.reset_index(drop=True)).reset_index(drop=True)

# Añadir 1 al índice para que empiece en 1 en lugar de 0Users
df.index = df.groupby('label').cumcount() + 1

df['new_format'] = df['label'].astype(str).apply(lambda x: x.ljust(5)) + "Q0".ljust(5) + df['docid'].apply(lambda x: x.ljust(15)) + df.index.astype(str).str.zfill(4)+ "    " + round(df['relevance'], 2).astype(str).apply(lambda x: x.ljust(8)) + "APB-UC3M_sentsim-all-MiniLM-L6-v2"

# Guardar el nuevo formato en un archivo txtAPB-UC3M_sentsim-all-MiniLM-L6-v2
df['new_format'].to_csv('submission/all-MiniLM-L6-v2.txt', index=False, header=False)


df = pd.read_csv('submission/all-MiniLM-L12-v2.csv')
# Supongamos que 'df' es tu dataframe final
df = df.groupby('label').apply(lambda x: x.reset_index(drop=True)).reset_index(drop=True)

# Añadir 1 al índice para que empiece en 1 en lugar de 0
df.index = df.groupby('label').cumcount() + 1

df['new_format'] = df['label'].astype(str).apply(lambda x: x.ljust(5)) + "Q0".ljust(5) + df['docid'].apply(lambda x: x.ljust(15)) + df.index.astype(str).str.zfill(4)+ "    " + round(df['relevance'], 2).astype(str).apply(lambda x: x.ljust(8)) + "APB-UC3M_sentsim-all-MiniLM-L12-v2"

# Guardar el nuevo formato en un archivo txtAPB-UC3M_sentsim-all-MiniLM-L6-v2
df['new_format'].to_csv('submission/all-MiniLM-L12-v2.txt', index=False, header=False)


df = pd.read_csv('submission/all-mpnet-base-v2.csv')
# Supongamos que 'df' es tu dataframe final
df = df.groupby('label').apply(lambda x: x.reset_index(drop=True)).reset_index(drop=True)

# Añadir 1 al índice para que empiece en 1 en lugar de 0
df.index = df.groupby('label').cumcount() + 1

df['new_format'] = df['label'].astype(str).apply(lambda x: x.ljust(5)) + "Q0".ljust(5) + df['docid'].apply(lambda x: x.ljust(15)) + df.index.astype(str).str.zfill(4)+ "    " + round(df['relevance'], 2).astype(str).apply(lambda x: x.ljust(8)) + "APB-UC3M_sentsim-all-mpnet-base-v2"

# Guardar el nuevo formato en un archivo txtAPB-UC3M_sentsim-all-MiniLM-L6-v2
df['new_format'].to_csv('submission/all-mpnet-base-v2.txt', index=False, header=False)


df = pd.read_csv('submission/roberta-base-go_emotions.csv')
# Supongamos que 'df' es tu dataframe final
df = df.groupby('label').apply(lambda x: x.reset_index(drop=True)).reset_index(drop=True)

# Añadir 1 al índice para que empiece en 1 en lugar de 0Users
df.index = df.groupby('label').cumcount() + 1

df['new_format'] = df['label'].astype(str).apply(lambda x: x.ljust(5)) + "Q0".ljust(5) + df['docid'].apply(lambda x: x.ljust(15)) + df.index.astype(str).str.zfill(4)+ "    " + round(df['relevance'], 2).astype(str).apply(lambda x: x.ljust(8)) + "APB-UC3M_classifier_roberta-base-go_emotions"

# Guardar el nuevo formato en un archivo txtAPB-UC3M_sentsim-all-MiniLM-L6-v2
df['new_format'].to_csv('submission/roberta-base-go_emotions.txt', index=False, header=False)


df = pd.read_csv('ensemble_final.csv')

# Ordenar por 'label' y 'relevance' de manera descendente
df.sort_values(['label', 'relevance'], ascending=[True, False], inplace=True)

# Seleccionar los primeros 1000 registros de cada 'label'
df = df.groupby('label').head(1000)

# Guardar el nuevo dataframe en un csv
df.to_csv('ensemble_final_final.csv', index=False)

df = pd.read_csv('ensemble_final_final.csv')
# Supongamos que 'df' es tu dataframe final
df = df.groupby('label').apply(lambda x: x.reset_index(drop=True)).reset_index(drop=True)

# Añadir 1 al índice para que empiece en 1 en lugar de 0Users
df.index = df.groupby('label').cumcount() + 1

df['new_format'] = df['label'].astype(str).apply(lambda x: x.ljust(5)) + "Q0".ljust(5) + df['docid'].apply(lambda x: x.ljust(15)) + df.index.astype(str).str.zfill(4)+ "    " + round(df['relevance'], 2).astype(str).apply(lambda x: x.ljust(8)) + "APB-UC3M_ensemble"

# Guardar el nuevo formato en un archivo txtAPB-UC3M_sentsim-all-MiniLM-L6-v2
df['new_format'].to_csv('submission/results5.txt', index=False, header=False)
